servicio_base_de_datos/requirements.txt
paho-mqtt
influxdb-client
python-dotenv